Address more review comments and CI failures.
